FIX: 집 들어가기 API 위치 변경사항 테스트에도 반영
